+++ src/App.jsx
@@ -1,4 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import { ExamProvider } from './context/ExamContext'
 import Layout from './components/Layout/Layout'
 import Home from './pages/Home'
 import ExamConfig from './pages/ExamConfig'
@@ -7,16 +8,18 @@ import ResultsPage from './pages/ResultsPage'
 
 function App() {
   return (
-    <Router>
-      <Layout>
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/exam/:certId/config" element={<ExamConfig />} />
-          <Route path="/exam/:certId/start" element={<ExamPage />} />
-          <Route path="/exam/:certId/results" element={<ResultsPage />} />
-        </Routes>
-      </Layout>
-    </Router>
+    <ExamProvider>
+      <Router>
+        <Layout>
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/exam/:certId/config" element={<ExamConfig />} />
+            <Route path="/exam/:certId/start" element={<ExamPage />} />
+            <Route path="/exam/:certId/results" element={<ResultsPage />} />
+          </Routes>
+        </Layout>
+      </Router>
+    </ExamProvider>
   )
 }